refactor auth handling
